Exception handling for REST controllers.

+++ b2w-starwars-api/src/main/java/br/com/b2w/starwars/api/service/PlanetService.java
@@ -1,10 +1,12 @@
 package br.com.b2w.starwars.api.service;
 
-import br.com.b2w.starwars.api.domain.Planet;
-import br.com.b2w.starwars.api.repository.PlanetRepository;
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import br.com.b2w.starwars.api.domain.Planet;
+import br.com.b2w.starwars.api.exceptions.PlanetNotFoundException;
+import br.com.b2w.starwars.api.repository.PlanetRepository;
 
 @Service
 public class PlanetService {
@@ -26,15 +28,22 @@ public class PlanetService {
 		return ! planetRepo.exists(planetId);
 	}
 
-	public Planet getPlanet(String planetId) {
+	public Planet getPlanet(String planetId) throws PlanetNotFoundException {
+		if (notExists(planetId)) {
+			throw new PlanetNotFoundException("Planet do not exists with ID: " + planetId);
+		}
 		return planetRepo.findOne(planetId);
 	}
 
-	public Planet searchPlanet(String name) {
-		return planetRepo.findByName(name);
+	public Planet searchPlanet(String name) throws PlanetNotFoundException {
+		return planetRepo.findByName(name)
+						.orElseThrow(() -> new PlanetNotFoundException("Planet not found with name: " + name));
 	}
 
-	public void remove(String planetId) {
+	public void remove(String planetId) throws PlanetNotFoundException {
+		if (notExists(planetId)) {
+			throw new PlanetNotFoundException("Planet do not exists with ID: " + planetId);
+		}
 		planetRepo.delete(planetId);		
 	}
